code/requirements.txt
h5py
numpy
PyYAML
scikit-image
scikit-learn
scipy
six
tensorflow
tensorboard
tensorboardX
torch==1.5.0
torchvision==0.6.0
colorful
colorlog
Pillow==6.1
pandas
openpyxl
xlrd
